+++ sinks/cortex/cortex_test.go
@@ -2,6 +2,7 @@ package cortex
 
 import (
 	"context"
+	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"io"
@@ -21,12 +22,13 @@ import (
 	"github.com/stripe/veneur/v14/samplers"
 	"github.com/stripe/veneur/v14/sinks"
 	"github.com/stripe/veneur/v14/trace"
+	"github.com/stripe/veneur/v14/util"
 )
 
 func TestName(t *testing.T) {
 	// Implicitly test that CortexMetricsSink implements MetricSink
 	var sink sinks.MetricSink
-	sink, err := NewCortexMetricSink("https://localhost/", 30, "", logrus.NewEntry(logrus.New()), "cortex", map[string]string{})
+	sink, err := NewCortexMetricSink("https://localhost/", 30, "", logrus.NewEntry(logrus.New()), "cortex", map[string]string{}, map[string]string{}, nil)
 
 	assert.NoError(t, err)
 	assert.Equal(t, "cortex", sink.Name())
@@ -38,7 +40,7 @@ func TestFlush(t *testing.T) {
 	defer server.Close()
 
 	// Set up a sink
-	sink, err := NewCortexMetricSink(server.URL, 30*time.Second, "", logrus.NewEntry(logrus.New()), "test", map[string]string{"corge": "grault"})
+	sink, err := NewCortexMetricSink(server.URL, 30*time.Second, "", logrus.NewEntry(logrus.New()), "test", map[string]string{"corge": "grault"}, map[string]string{}, nil)
 	assert.NoError(t, err)
 	assert.NoError(t, sink.Start(trace.DefaultClient))
 
@@ -52,19 +54,21 @@ func TestFlush(t *testing.T) {
 	assert.NoError(t, sink.Flush(context.Background(), metrics))
 
 	// Retrieve the data which the server received
-	data, err := server.Latest()
+	data, headers, err := server.Latest()
 	assert.NoError(t, err)
+
+	// Check standard headers
+	assert.True(t, hasHeader(*headers, "Content-Encoding", "snappy"), "missing required Content-Encoding header")
+	assert.True(t, hasHeader(*headers, "Content-Type", "application/x-protobuf"), "missing required Content-Type header")
+	assert.True(t, hasHeader(*headers, "User-Agent", "veneur/cortex"), "missing required User-Agent header")
+	assert.True(t, hasHeader(*headers, "X-Prometheus-Remote-Write-Version", "0.1.0"), "missing required version header")
 
 	// The underlying method to convert metric -> timeseries does not
 	// preserve order, so we're sorting the data here
 	for k := range data.Timeseries {
 		sort.Slice(data.Timeseries[k].Labels, func(i, j int) bool {
 			val := strings.Compare(data.Timeseries[k].Labels[i].Name, data.Timeseries[k].Labels[j].Name)
-			if val == -1 {
-				return true
-			}
-
-			return false
+			return val == -1
 		})
 
 	}
@@ -79,10 +83,168 @@ func TestFlush(t *testing.T) {
 	assert.Equal(t, string(expected), string(actual))
 }
 
+func TestCustomHeaders(t *testing.T) {
+	// Listen for prometheus writes
+	server := NewTestServer(t)
+	defer server.Close()
+
+	// Define custom headers
+	customHeaders := map[string]string{
+		"Authorization":    "Bearer 12345",
+		"My-Custom-Header": "testing-123",
+		"Another-Header":   "foobar",
+	}
+
+	// Set up a sink with custom headers
+	sink, err := NewCortexMetricSink(server.URL, 30*time.Second, "", logrus.NewEntry(logrus.New()), "test", map[string]string{"corge": "grault"}, customHeaders, nil)
+	assert.NoError(t, err)
+	assert.NoError(t, sink.Start(trace.DefaultClient))
+
+	// input.json contains three timeseries samples in InterMetrics format
+	jsInput, err := ioutil.ReadFile("testdata/input.json")
+	assert.NoError(t, err)
+	var metrics []samplers.InterMetric
+	assert.NoError(t, json.Unmarshal(jsInput, &metrics))
+
+	// Perform the flush to the test server
+	assert.NoError(t, sink.Flush(context.Background(), metrics))
+
+	// Retrieve the headers which the server received
+	_, headers, err := server.Latest()
+	assert.NoError(t, err)
+
+	// Check custom headers
+	for name, value := range customHeaders {
+		assert.True(t, hasHeader(*headers, name, value), "Missing header "+name)
+	}
+}
+
+func TestBasicAuth(t *testing.T) {
+	// Listen for prometheus writes
+	server := NewTestServer(t)
+	defer server.Close()
+
+	// Define custom headers
+	customHeaders := map[string]string{
+		"My-Custom-Header": "testing-456",
+		"Another-Header":   "bazzoo",
+	}
+	auth := BasicAuthType{
+		Username: util.StringSecret{Value: "user1"},
+		Password: util.StringSecret{Value: "p@ssWerd"},
+	}
+
+	// Set up a sink with custom headers
+	sink, err := NewCortexMetricSink(server.URL, 30*time.Second, "", logrus.NewEntry(logrus.New()), "test", map[string]string{"corge": "grault"}, customHeaders, &auth)
+	assert.NoError(t, err)
+	assert.NoError(t, sink.Start(trace.DefaultClient))
+
+	// input.json contains three timeseries samples in InterMetrics format
+	jsInput, err := ioutil.ReadFile("testdata/input.json")
+	assert.NoError(t, err)
+	var metrics []samplers.InterMetric
+	assert.NoError(t, json.Unmarshal(jsInput, &metrics))
+
+	// Perform the flush to the test server
+	assert.NoError(t, sink.Flush(context.Background(), metrics))
+
+	// Retrieve the headers which the server received
+	_, headers, err := server.Latest()
+	assert.NoError(t, err)
+
+	// Check custom headers
+	for name, value := range customHeaders {
+		assert.True(t, hasHeader(*headers, name, value), "Missing or incorrect "+name+" header")
+	}
+	authString := auth.Username.Value + ":" + auth.Password.Value
+	assert.True(t, hasHeader(*headers, "Authorization", "Basic "+base64.StdEncoding.EncodeToString([]byte(authString))),
+		"Missing or invalid Authorization header")
+}
+
+func TestParseConfig(t *testing.T) {
+	testConfigValues := map[string]interface{}{
+		"url":            "this://is.a.url",
+		"remote_timeout": "90s",
+		"proxy_url":      "http://another.url:8000",
+		"headers":        map[string]string{"My-Header": "a-header-value"},
+		"authorization": map[string]interface{}{
+			"credentials": "the-credential",
+		},
+	}
+
+	parsedConfig, err := ParseConfig("cortex", testConfigValues)
+	assert.NoError(t, err)
+	cortexConfig := parsedConfig.(CortexMetricSinkConfig)
+	assert.Equal(t, cortexConfig.URL, testConfigValues["url"])
+	assert.Equal(t, cortexConfig.RemoteTimeout, time.Duration(90*time.Second))
+	assert.Equal(t, cortexConfig.ProxyURL, testConfigValues["proxy_url"])
+	assert.Equal(t, cortexConfig.Headers, testConfigValues["headers"])
+	assert.NotNil(t, cortexConfig.Authorization)
+	assert.Equal(t, cortexConfig.Authorization.Type, DefaultAuthorizationType)
+	assert.Equal(t, cortexConfig.Authorization.Credential.Value, "the-credential")
+	assert.Empty(t, cortexConfig.BasicAuth)
+}
+
+func TestParseConfigBasicAuth(t *testing.T) {
+	testConfigValues := map[string]interface{}{
+		"url":            "this://is.a.url",
+		"remote_timeout": "90s",
+		"proxy_url":      "http://another.url:8000",
+		"basic_auth": map[string]interface{}{
+			"username": "user",
+			"password": "pwd",
+		},
+	}
+
+	parsedConfig, err := ParseConfig("cortex", testConfigValues)
+	assert.NoError(t, err)
+	cortexConfig := parsedConfig.(CortexMetricSinkConfig)
+	assert.Equal(t, cortexConfig.URL, testConfigValues["url"])
+	assert.Equal(t, cortexConfig.RemoteTimeout, time.Duration(90*time.Second))
+	assert.Equal(t, cortexConfig.ProxyURL, testConfigValues["proxy_url"])
+	assert.Empty(t, cortexConfig.Headers)
+	assert.Empty(t, cortexConfig.Authorization)
+	assert.NotNil(t, cortexConfig.BasicAuth)
+	assert.Equal(t, cortexConfig.BasicAuth.Username.Value, "user")
+	assert.Equal(t, cortexConfig.BasicAuth.Password.Value, "pwd")
+}
+
+func TestParseConfigDuplicateAuth(t *testing.T) {
+	testConfigValues := map[string]interface{}{
+		"url":            "this://is.a.url",
+		"remote_timeout": "90s",
+		"proxy_url":      "http://another.url:8000",
+		"basic_auth": map[string]interface{}{
+			"username": "user",
+			"password": "pwd",
+		},
+		"authorization": map[string]interface{}{
+			"credentials": "the-credential",
+		},
+	}
+
+	_, err := ParseConfig("cortex", testConfigValues)
+	assert.Error(t, err)
+}
+
+func TestParseConfigBadBasicAuth(t *testing.T) {
+	testConfigValues := map[string]interface{}{
+		"url":            "this://is.a.url",
+		"remote_timeout": "90s",
+		"proxy_url":      "http://another.url:8000",
+		"basic_auth": map[string]interface{}{
+			"username": "user",
+		},
+	}
+
+	_, err := ParseConfig("cortex", testConfigValues)
+	assert.Error(t, err)
+}
+
 func TestCorrectlySetTimeout(t *testing.T) {
 	timeouts := []int{10, 20, 30, 17, 21}
 	for to := range timeouts {
-		sink, err := NewCortexMetricSink("http://noop", time.Duration(to), "", logrus.NewEntry(logrus.New()), "test", map[string]string{"corge": "grault"})
+		sink, err := NewCortexMetricSink("http://noop", time.Duration(to), "", logrus.NewEntry(logrus.New()), "test", map[string]string{"corge": "grault"}, map[string]string{}, nil)
 		assert.NoError(t, err)
 
 		err = sink.Start(&trace.Client{})
@@ -149,9 +311,10 @@ func BenchmarkSanitise(b *testing.B) {
 // TestServer wraps an internal httptest.Server and provides a convenience
 // method for retrieving the most recently written series
 type TestServer struct {
-	URL    string
-	data   *prompb.WriteRequest
-	server *httptest.Server
+	URL     string
+	headers *http.Header
+	data    *prompb.WriteRequest
+	server  *httptest.Server
 }
 
 // Close closes the internal test server
@@ -160,11 +323,11 @@ func (t *TestServer) Close() {
 }
 
 // Latest returns the most recent write request, or errors if there was none
-func (t *TestServer) Latest() (*prompb.WriteRequest, error) {
+func (t *TestServer) Latest() (*prompb.WriteRequest, *http.Header, error) {
 	if t.data == nil {
-		return nil, errors.New("no data received")
+		return nil, nil, errors.New("no data received")
 	}
-	return t.data, nil
+	return t.data, t.headers, nil
 }
 
 // NewTestServer starts a test server instance. Ensure calls are followed by
@@ -188,7 +351,8 @@ func NewTestServer(t *testing.T) *TestServer {
 			http.Error(w, "missing headers", http.StatusBadRequest)
 			return
 		}
-		// keep a record of the most recently received request
+		// keep a record of the most recently received headers, request
+		result.headers = &r.Header
 		result.data = wr
 	})